+++ NEWS.md
@@ -1,3 +1,10 @@
+# hip-viz (dev)
+
+## Minor updates
+
+- Use `migbirdHIP:::assignFlyway()` for flyway assignments.
+- Use `{httr}` and `{jsonlite}` to get the most recent commit from the `hip-viz` repo to return "last updated" date to users, rather that previous method which is uninformative on live release (`lubridate::now()`).
+
 # hip-viz 0.1.0
 
 Launched with 2025-2026 Harvest Information Program registration summary data.